--- a/SimPEG/dask/simulation.py
+++ b/SimPEG/dask/simulation.py
@@ -1,6 +1,20 @@
 from ..simulation import BaseSimulation as Sim
 
-Sim._max_ram = 16
+
+@property
+def chunk_format(self):
+    "Apply memory chunks along rows of G, either 'equal', 'row', or 'auto'"
+    return self._chunk_format
+
+
+@chunk_format.setter
+def chunk_format(self, other):
+    if other not in ["equal", "row", "auto"]:
+        raise ValueError("Chunk format must be 'equal', 'row', or 'auto'")
+    self._chunk_format = other
+
+
+Sim.chunk_format = chunk_format
 
 
 @property
@@ -18,8 +32,6 @@ def max_ram(self, other):
 
 Sim.max_ram = max_ram
 
-Sim._max_chunk_size = 128
-
 
 @property
 def max_chunk_size(self):
@@ -35,3 +47,39 @@ def max_chunk_size(self, other):
 
 
 Sim.max_chunk_size = max_chunk_size
+
+# add dask options to BaseSimulation.__init__
+_old_init = Sim.__init__
+
+
+def __init__(
+    self,
+    mesh=None,
+    survey=None,
+    solver=None,
+    solver_opts=None,
+    sensitivity_path="./sensitivity/",
+    counter=None,
+    verbose=False,
+    chunk_format="row",
+    max_ram=16,
+    max_chunk_size=128,
+    **kwargs,
+):
+    _old_init(
+        self,
+        mesh=mesh,
+        survey=survey,
+        solver=solver,
+        solver_opts=solver_opts,
+        sensitivity_path=sensitivity_path,
+        counter=counter,
+        verbose=verbose,
+        **kwargs,
+    )
+    self.chunk_format = chunk_format
+    self.max_ram = max_ram
+    self.max_chunk_size = max_chunk_size
+
+
+Sim.__init__ = __init__