--- a/SimPEG/potential_fields/gravity/receivers.py
+++ b/SimPEG/potential_fields/gravity/receivers.py
@@ -1,4 +1,3 @@
-import numpy as np
 from ... import survey
 
 
@@ -6,54 +5,42 @@ class Point(survey.BaseRx):
     """
     Gravity point receiver class for integral formulation
 
+    Parameters
+    ----------
+    locations : (n_loc, dim) array_like
+    components : str or list of str
+
     :param numpy.ndarray locations: receiver locations index (ie. :code:`np.c_[ind_1, ind_2, ...]`)
     :param string component: receiver component
          "gx", "gy", "gz", "gxx", "gxy", "gxz",
-         "gyy", "gyz", "gzz", "guv", "amp" [default]
+         "gyy", "gyz", "gzz", "guv"
     """
 
     def __init__(self, locations, components="gz", **kwargs):
-
         super(survey.BaseRx, self).__init__(locations=locations, **kwargs)
-
-        n_locations = locations.shape[0]
-
         if isinstance(components, str):
             components = [components]
 
-        component_dict = {}
         for component in components:
-            component_dict[component] = np.ones(n_locations, dtype="bool")
+            if component not in [
+                "gx",
+                "gy",
+                "gz",
+                "gxx",
+                "gxy",
+                "gxz",
+                "gyy",
+                "gyz",
+                "gzz",
+                "guv",
+            ]:
+                raise ValueError(
+                    f"{component} not recognized, must be one of "
+                    "'gx', 'gy', 'gz', 'gxx', 'gxy', 'gxz'"
+                    "'gyy', 'gyz', 'gzz', or 'guv'"
+                )
+        self.components = components
 
-        assert np.all(
-            [
-                component
-                in ["gx", "gy", "gz", "gxx", "gxy", "gxz", "gyy", "gyz", "gzz", "guv"]
-                for component in list(component_dict.keys())
-            ]
-        ), (
-            "Components {0!s} not known. Components must be in "
-            "'gx', 'gy', 'gz', 'gxx', 'gxy', 'gxz'"
-            "'gyy', 'gyz', 'gzz', 'guv'"
-            "Arbitrary orientations have not yet been "
-            "implemented.".format(component)
-        )
-        self.components = component_dict
-
+    @property
     def nD(self):
-
-        if self.receiver_index is not None:
-
-            return self.location_index.shape[0]
-
-        elif self.locations is not None:
-
-            return self.locations.shape[0]
-
-        else:
-
-            return None
-
-    def receiver_index(self):
-
-        return self.receiver_index
+        return self.locations.shape[0] * len(self.components)