… .azure-pipelines/matrix.yml
… SimPEG/potential_fields/gravity/simulation.py
… SimPEG/potential_fields/gravity/survey.py
… SimPEG/potential_fields/magnetics/receivers.py
… SimPEG/potential_fields/magnetics/simulation.py
… SimPEG/potential_fields/magnetics/sources.py
… SimPEG/potential_fields/magnetics/survey.py
… tests/dask/test_grav_inversion_linear.py
… tests/dask/test_mag_MVI_Octree.py
… tests/dask/test_mag_inversion_linear_Octree.py
… tests/pf/test_forward_Grav_Linear.py
… tests/pf/test_forward_Mag_Linear.py
… tests/pf/test_grav_inversion_linear.py
… tests/pf/test_survey_counting.py
… tests/utils/test_io_utils.py