This PR overhalls the internals of the potential fields integral simulations to make use of kernels defined in `geoana`. There should be no visible change on the end user.

Things it does
* Uses the compiled kernels from geoana.
* Only calculates the kernel functions at unique nodes.
* Adds multiprocessing back as the default way of computing the sensitivity matrices