@jcapriot 2022-10-24
Some timing examples on my laptop (4 processors):
on a 32 x 32 x 32 mesh (32768 cells) with 101 * 101 = 10201 observation locations.

Before and after timings for creations of a sensitivity matrix:
Gravity: 
- single "gz" component:
   ~200 seconds -> ~13 seconds
- All 6 components of GG: "gxx, gxy, gxz, gyy, gyz, gzz"
   ~430 seconds -> ~43 seconds

Magnetics:
- "tmi"
  ~ 500 seconds -> ~ 30 seconds