@thibaut-kobold 2022-10-24
> Some timing examples on my laptop (4 processors): on a 32 x 32 x 32 mesh (32768 cells) with 101 * 101 = 10201 observation locations.
> 
> Before and after timings for creations of a sensitivity matrix: Gravity:
> 
> * single "gz" component:
>   ~200 seconds -> ~13 seconds
> * All 6 components of GG: "gxx, gxy, gxz, gyy, gyz, gzz"
>   ~430 seconds -> ~43 seconds
> 
> Magnetics:
> 
> * "tmi"
>   ~ 500 seconds -> ~ 30 seconds

Very cool, congrats!
Quick question: looking at the geoana kernels, I see [there is a pure-python version, and a C (or C++) version of the kernels](https://github.com/simpeg/geoana/blob/3d3e021a859670a675227aa274db6d158fbafe4d/geoana/kernels/potential_field_prism.py#L289) and that it tries to import the C version first and revert to pure python if it fails. I imagine the speed up is achieved by the C version My question: is there a way to display a warning if the pure-python version is imported instead of the accelerated one?