@jcapriot 2022-10-24
The speedup between the two different versions is not actually that significant. If you're installing geoana through conda-forge it will come compiled. I.e. the c one is maybe 80% the speed of the numpy one, if i remember correctly. Most of the timing differences here come from the multiprocessing addition, and the non-repeated calls of the functions at nodes.